Move federal code to utils (#3591)
* Move federal base code to utils

* Migrate federated.js

* Fix linting errors
