+++ libs/blocks/global-navigation/features/breadcrumbs/breadcrumbs.js
@@ -1,6 +1,5 @@
-import { getMetadata, getConfig } from '../../../../utils/utils.js';
+import { getMetadata, getConfig, getFederatedUrl } from '../../../../utils/utils.js';
 import { toFragment, lanaLog } from '../../utilities/utilities.js';
-import { getFederatedUrl } from '../../../../utils/federated.js';
 
 const metadata = {
   seo: 'breadcrumbs-seo',
