+++ libs/blocks/global-navigation/features/profile/dropdown.js
@@ -1,7 +1,6 @@
-import { getConfig } from '../../../../utils/utils.js';
+import { getConfig, getFedsPlaceholderConfig } from '../../../../utils/utils.js';
 import { toFragment, trigger, closeAllDropdowns, logErrorFor } from '../../utilities/utilities.js';
 import { replaceKeyArray } from '../../../../features/placeholders.js';
-import { getFedsPlaceholderConfig } from '../../../../utils/federated.js';
 
 const getLanguage = (ietfLocale) => {
   if (!ietfLocale.length) return 'en';
