+++ libs/features/personalization/personalization.js
@@ -2,8 +2,14 @@
 /* eslint-disable no-underscore-dangle */
 /* eslint-disable no-console */
 
-import { createTag, getConfig, loadLink, loadScript, localizeLink } from '../../utils/utils.js';
-import { getFederatedUrl } from '../../utils/federated.js';
+import {
+  createTag,
+  getConfig,
+  loadLink,
+  loadScript,
+  localizeLink,
+  getFederatedUrl,
+} from '../../utils/utils.js';
 
 /* c8 ignore start */
 const PHONE_SIZE = window.screen.width < 550 || window.screen.height < 550;
