+++ test/features/georoutingv2/georoutingv2.test.js
@@ -4,8 +4,7 @@ import { setViewport } from '@web/test-runner-commands';
 
 const { default: init, getCookie } = await import('../../../libs/features/georoutingv2/georoutingv2.js');
 let { getMetadata } = await import('../../../libs/utils/utils.js');
-const { getFederatedContentRoot } = await import('../../../libs/utils/federated.js');
-const { createTag, loadStyle, loadBlock, setConfig } = await import('../../../libs/utils/utils.js');
+const { createTag, loadStyle, loadBlock, setConfig, getFederatedContentRoot } = await import('../../../libs/utils/utils.js');
 
 const mockConfig = {
   locales: {
