+++ test/utils/federated.test.js
@@ -1,5 +1,5 @@
 import { expect } from '@esm-bundle/chai';
-import { getFederatedUrl, getFederatedContentRoot } from '../../libs/utils/federated.js';
+import { getFederatedUrl, getFederatedContentRoot } from '../../libs/utils/utils.js';
 
 const baseHost = 'https://main--federal--adobecom.aem.page';
